+++ geolocator_android/android/src/main/java/com/baseflow/geolocator/location/LocationAccuracyStatus.java
@@ -6,8 +6,4 @@ public enum LocationAccuracyStatus {
 
   /// The precise location of the device will be returned.
   precise,
-
-  /// When an 'unknown' status is returned, Location Permission is not yet requested or denied
-  // (forever).
-  unknown,
 }